+++ config.go
@@ -48,6 +48,7 @@ type Config struct {
 	UseAsyncSessionWrite bool `json:"optimisations_use_async_session_write"`
     AllowMasterKeys bool `json:"allow_master_keys"`
     HashKeys bool `json:"hash_keys"`
+    SuppressRedisSignalReload bool `json:"suppress_redis_signal_reload"`
 }
 
 // WriteDefaultConf will create a default configuration file and set the storage type to "memory"